Fix: WC2-905: Improve ETL for WFP DHIS2 exports
Refs: #2818 [WC2-905]
* Map monthly statistics table to data to push to dhis2

* Getting and aggregate journeys data by org unit and period

* Aggregate journey data by period(month and year) and org unit

* Restructuring and map monthly statistics analytic table to data to push to dhis2

* Refactor query to aggregate journey data to monthly statistics

* Indexing needed fields on analytics table to make aggregate query run faster

* Refactor query to aggregate data monthly by org unit and period

* Fix aggregating fields based in program_type

* Show entry point for visit

* WIP: Aggregate data to push to dhis2

* Indexing some fields on analytics tables and restructure monthly statistics table to match with the data to push to dhis2

* Show oedema in visit analytic table

* Exclude non eligible montly statistics

* Remap dhis2 dataelement with the fields from monthly statistics table

* Refactor functions to aggregate data to push to dhis2

* Paginate query to aggregate monthly data per org unit and month

* Check if the visit has oedema status

* Calculating cases referred from community health worker

* Adding needed fields to map the monthly statistics table with the data to push to dhis2

* Apply ruff formater

* Getting screening data and cases from community health worker

* Fix migration

* Fix aggregate data to push to dhis2

* Apply ruff

* Update translations

* Remove unused code

* Remove unused fields

* Fix ETL tests

* Reuse refactored v2 on Ethiopia and Nigeria

* Move screening data query in the v2 common file

* Apply ruff formater

* Remove no needed v1 code and rename v2 to common file

* Generate monthly statistics for only org units related to newly created/updated data

* Allow to clean analytics data related to an account

* Push data for only org units related to newly created/updated data in a period

* Fix dataelement in the mapper file and convert to integer the screening data to push to dhis2

* Remove all ETL remaining old code

* Show in log the number of deleted itesms

* Fix aggregating screening data to push to dhis2

* Fix community health worker categoryOptionCombo for PBWG

* Fix screening reporting and community health worker data

* Fix fields name for community health worker

* Try to avoid f-string when showing the number of cleaned data

* Use print instead of logger when showing the number of deleted items

* Check first physiology status before the gender

* Allow to filter beneficiaries by physiology status from django admin

* Show the deletion summary when cleaning all analytics data

* Fix PBWG org_units and aggregation date extraction

- Use PBWG-specific org_units for PBWG monthly statistics reset and
  aggregation in etl_ng, etl_ssd, and etl_ethiopia tasks instead of
  reusing the U5 org_units, which would reset/aggregate the wrong data.
- Fix _retrieve_aggregated_journeys_data to extract year/month from the
  visit's own date field instead of journey__visit__date, which traverses
  back through the reverse FK to all visits and causes cross-joins.

* Fix missing account filter when deleting Ethiopia PBWG beneficiaries

Without the account filter, the delete could affect beneficiaries from
other accounts that happen to share entity IDs.

* Pass account_id to delete_beneficiaries_analytics template

The view now takes account_id as a URL parameter but wasn't passing it
to the template context. Show which account will be affected in the
confirmation page.

* Fix UnorderedObjectListWarning in ETL pagination

Add order_by to get_updated_entity_ids so the Paginator produces
consistent pages. Without ordering, entities could be skipped or
duplicated across pages.

---------
