+++ plugins/wfp/locale/fr/LC_MESSAGES/django.po
@@ -7,7 +7,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: \n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-12-10 15:51+0000\n"
+"POT-Creation-Date: 2026-03-17 08:48+0000\n"
 "PO-Revision-Date: 2025-02-05 14:41+0100\n"
 "Last-Translator: Olivier Le Thanh Duong <olivier@lethanh.be>\n"
 "Language-Team: \n"
@@ -89,6 +89,10 @@ msgstr "OTP - Under 6"
 #: plugins/wfp/models.py
 msgid "BSFP"
 msgstr "BSFP"
+
+#: plugins/wfp/models.py
+msgid "Not Eligible"
+msgstr "Non Eligible"
 
 #: plugins/wfp/models.py
 msgid "Breastfeeding"
