+++ plugins/wfp/tasks.py
@@ -7,13 +7,9 @@ from django.db import connection
 from iaso.management.commands import unique_indexes
 from iaso.management.commands.clean_up_duplicate_submissions import DRY_RUN_ARG
 from iaso.models.base import ExternalCredentials
+from plugins.wfp.aggregator import Aggregator
 from plugins.wfp.common import ETL
 
-from .management.commands.ethiopia.Under5 import ET_Under5
-from .management.commands.nigeria.Pbwg import NG_PBWG
-from .management.commands.nigeria.Under5 import NG_Under5
-from .management.commands.south_sudan.Dhis2 import Dhis2
-from .management.commands.south_sudan.Screening import Screening
 from .models import *
 
 
@@ -81,6 +77,9 @@ def etl_ng(all_data=None):
     """Extract beneficiary data from Iaso tables and store them in the format expected by existing tableau dashboards"""
     from django_celery_results.models import TaskResult
 
+    from .management.commands.nigeria.Pbwg import NG_PBWG
+    from .management.commands.nigeria.Under5 import NG_Under5
+
     task_name = "plugins.wfp.tasks.etl_ng"
     last_success_task = TaskResult.objects.filter(task_name=task_name, status="SUCCESS").first()
 
@@ -99,37 +98,29 @@ def etl_ng(all_data=None):
     logger.info("Starting ETL for Nigeria")
     entity_type_U5_code = "nigeria_under5"
     etl_u5 = ETL(entity_type_U5_code)
-    etl = ETL()
-    account = etl_u5.account_related_to_entity_type()
+    account = etl_u5.get_account()
     updated_U5_beneficiaries = etl_u5.get_updated_entity_ids(last_success_task_date)
     Beneficiary.objects.filter(account=account, entity_id__in=updated_U5_beneficiaries).delete()
-
-    NG_Under5().run(entity_type_U5_code, updated_U5_beneficiaries, task_name)
+    NG_Under5().run(updated_U5_beneficiaries, entity_type_U5_code, task_name)
     logger.info(
         f"----------------------------- Aggregating journey for {account} per org unit, admission and period(month and year) -----------------------------"
     )
-    org_units_with_updated_data = etl_u5.get_org_unit_ids_with_updated_data(last_success_task_date)
-    MonthlyStatistics.objects.filter(
-        account=account,
-        programme_type="U5",
-        org_unit_id__in=org_units_with_updated_data,
-    ).delete()
-    etl.journey_with_visit_and_steps_per_visit(account, "U5", org_units_with_updated_data)
+    org_units = etl_u5.get_org_unit_and_period_with_updated_data(last_success_task_date)
+    Aggregator.reset_monthly_statistics(account, "U5", org_units)
+    Aggregator.aggregate_monthly_data_by_org_unit(account, org_units, "U5")
+
     entity_type_pbwg_code = "nigeria_pbwg"
     etl_pbwg = ETL(entity_type_pbwg_code)
-    pbwg_account = etl_pbwg.account_related_to_entity_type()
+    pbwg_account = etl_pbwg.get_account()
     updated_pbwg_beneficiaries = etl_pbwg.get_updated_entity_ids(last_success_task_date)
-    Beneficiary.objects.filter(entity_id__in=updated_pbwg_beneficiaries).delete()
-    NG_PBWG().run(entity_type_pbwg_code, updated_pbwg_beneficiaries, task_name)
+    Beneficiary.objects.filter(account=pbwg_account, entity_id__in=updated_pbwg_beneficiaries).delete()
+    NG_PBWG().run(updated_pbwg_beneficiaries, entity_type_pbwg_code, task_name)
     logger.info(
         f"----------------------------- Aggregating PBWG journey for {pbwg_account} per org unit, admission and period(month and year) -----------------------------"
     )
-    MonthlyStatistics.objects.filter(
-        account=pbwg_account,
-        programme_type="PLW",
-        org_unit_id__in=org_units_with_updated_data,
-    ).delete()
-    etl.journey_with_visit_and_steps_per_visit(pbwg_account, "PLW", org_units_with_updated_data)
+    pbwg_org_units = etl_pbwg.get_org_unit_and_period_with_updated_data(last_success_task_date)
+    Aggregator.reset_monthly_statistics(pbwg_account, "PLW", pbwg_org_units)
+    Aggregator.aggregate_monthly_data_by_org_unit(pbwg_account, pbwg_org_units, "PLW")
 
 
 @shared_task()
@@ -150,9 +141,8 @@ def ssd_aggregate_and_push_data_to_dhis2(all_data=None):
         last_success_task_date = None
 
     etl = ETL("ssd_under5")
-    account = etl.account_related_to_entity_type()
-    org_units_with_updated_data = etl.get_org_unit_ids_with_updated_data(last_success_task_date)
-
+    account = etl.get_account()
+    org_units = etl.get_org_unit_and_period_with_updated_data(last_success_task_date)
     external_credential = ExternalCredentials.objects.filter(account=account).first()
     if external_credential is not None and (
         external_credential.url is not None
@@ -160,13 +150,9 @@ def ssd_aggregate_and_push_data_to_dhis2(all_data=None):
         and external_credential.password is not None
     ):
         logger.info(
-            f"----------------------------- Aggregating monthly data to push to DHIS2 for {len(org_units_with_updated_data)} org unit on {account} -----------------------------"
+            f"----------------------------- Aggregating monthly data to push to DHIS2 for {account} -----------------------------"
         )
-        monthly_data = etl.aggregating_data_to_push_to_dhis2(account, org_units_with_updated_data)
-        pushed_data = Dhis2().save_dhis2_sync_results(external_credential, account, monthly_data)
-        logger.info(
-            f"----------------------------- Pushed to DHIS2 on U5 and PBW for {len(pushed_data)} rows aggregated per year and month -----------------------------"
-        )
+        Aggregator().aggregate_by_nutrition_program(account, org_units, external_credential)
     else:
         logger.info(
             f"----------------------------- No DHIS2 credentials found for {account} -----------------------------"
@@ -175,7 +161,7 @@ def ssd_aggregate_and_push_data_to_dhis2(all_data=None):
 
 @shared_task()
 def etl_ssd(all_data=None):
-    """ETL v2 for South Sudan Under-5 children and PBWG.
+    """ETL for South Sudan Under-5 children and PBWG.
 
     This is a rewrite of the Under-5 part of etl_ssd with key fixes:
     - Supports unlimited journeys per beneficiary (v1 limited to 2)
@@ -187,6 +173,7 @@ def etl_ssd(all_data=None):
     from django_celery_results.models import TaskResult
 
     from .management.commands.south_sudan.Pbwg import Pbwg
+    from .management.commands.south_sudan.Screening import Screening
     from .management.commands.south_sudan.Under5 import Under5
 
     task_name = "plugins.wfp.tasks.etl_ssd"
@@ -200,40 +187,32 @@ def etl_ssd(all_data=None):
     if all_data is not None:
         last_success_task_date = None
 
-    logger.info("Starting ETL v2 for South Sudan")
+    logger.info("Starting ETL for South Sudan")
 
     entity_type_u5_code = "ssd_under5"
     etl_u5 = ETL(entity_type_u5_code)
-    child_account = etl_u5.account_related_to_entity_type()
+    child_account = etl_u5.get_account()
     updated_beneficiaries = etl_u5.get_updated_entity_ids(last_success_task_date)
     Beneficiary.objects.filter(account=child_account, entity_id__in=updated_beneficiaries).delete()
 
     Under5().run(updated_beneficiaries, entity_type_u5_code, task_name)
 
-    logger.info(f"Aggregating Children under 5 journey (v2) for {child_account} per org unit, admission and period")
-    org_units_with_updated_data = etl_u5.get_org_unit_ids_with_updated_data(last_success_task_date)
-    MonthlyStatistics.objects.filter(
-        account=child_account,
-        programme_type="U5",
-        org_unit_id__in=org_units_with_updated_data,
-    ).delete()
-    etl_u5.journey_with_visit_and_steps_per_visit(child_account, "U5", org_units_with_updated_data)
+    logger.info(f"Aggregating Children under 5 journey for {child_account} per org unit, admission and period")
+    org_units = etl_u5.get_org_unit_and_period_with_updated_data(last_success_task_date)
+    Aggregator.reset_monthly_statistics(child_account, "U5", org_units)
+    Aggregator.aggregate_monthly_data_by_org_unit(child_account, org_units, "U5")
 
     entity_type_pbwg_code = "ssd_pbwg"
     etl_pbwg = ETL(entity_type_pbwg_code)
-    pbwg_account = etl_pbwg.account_related_to_entity_type()
+    pbwg_account = etl_pbwg.get_account()
     updated_pbwg_beneficiaries = etl_pbwg.get_updated_entity_ids(last_success_task_date)
     Beneficiary.objects.filter(account=pbwg_account, entity_id__in=updated_pbwg_beneficiaries).delete()
     Pbwg().run(updated_pbwg_beneficiaries, entity_type_pbwg_code, task_name)
 
-    logger.info(f"Aggregating PBWG journey (v2) for {pbwg_account} per org unit, admission and period")
-    org_units_with_updated_data = etl_pbwg.get_org_unit_ids_with_updated_data(last_success_task_date)
-    MonthlyStatistics.objects.filter(
-        account=pbwg_account,
-        programme_type="PLW",
-        org_unit_id__in=org_units_with_updated_data,
-    ).delete()
-    etl_pbwg.journey_with_visit_and_steps_per_visit(pbwg_account, "PLW", org_units_with_updated_data)
+    logger.info(f"Aggregating PBWG journey for {pbwg_account} per org unit, admission and period")
+    pbwg_org_units = etl_pbwg.get_org_unit_and_period_with_updated_data(last_success_task_date)
+    Aggregator.reset_monthly_statistics(pbwg_account, "PLW", pbwg_org_units)
+    Aggregator.aggregate_monthly_data_by_org_unit(pbwg_account, pbwg_org_units, "PLW")
 
     Screening().run(child_account, last_success_task_date)
 
@@ -243,6 +222,7 @@ def etl_ethiopia(all_data=None):
     """Extract beneficiary data from Iaso tables and store them in the format expected by existing tableau dashboards"""
     from django_celery_results.models import TaskResult
 
+    from .management.commands.ethiopia.Under5 import ET_Under5
     from .management.commands.south_sudan.Pbwg import Pbwg
 
     task_name = "plugins.wfp.tasks.etl_ethiopia"
@@ -263,36 +243,28 @@ def etl_ethiopia(all_data=None):
     logger.info("Starting ETL for Ethiopia")
     entity_type_U5_code = "ethiopia_under5"
     etl_u5 = ETL(entity_type_U5_code)
-    etl = ETL()
-    child_account = etl_u5.account_related_to_entity_type()
+    child_account = etl_u5.get_account()
     updated_U5_beneficiaries = etl_u5.get_updated_entity_ids(last_success_task_date)
     Beneficiary.objects.filter(account=child_account, entity_id__in=updated_U5_beneficiaries).delete()
-    ET_Under5().run(entity_type_U5_code, updated_U5_beneficiaries, task_name)
+    ET_Under5().run(updated_U5_beneficiaries, entity_type_U5_code, task_name)
 
     logger.info(
         f"----------------------------- Aggregating Children under 5 journey for {child_account} per org unit, admission and period(month and year) -----------------------------"
     )
-    org_units_with_updated_data = etl_u5.get_org_unit_ids_with_updated_data(last_success_task_date)
-    MonthlyStatistics.objects.filter(
-        account=child_account,
-        programme_type="U5",
-        org_unit_id__in=org_units_with_updated_data,
-    ).delete()
-    etl.journey_with_visit_and_steps_per_visit(child_account, "U5", org_units_with_updated_data)
+    org_units = etl_u5.get_org_unit_and_period_with_updated_data(last_success_task_date)
+    Aggregator.reset_monthly_statistics(child_account, "U5", org_units)
+    Aggregator.aggregate_monthly_data_by_org_unit(child_account, org_units, "U5")
 
     entity_type_pbwg_code = "ethiopia_pbwg"
     etl_pbwg = ETL(entity_type_pbwg_code)
-    pbwg_account = etl_pbwg.account_related_to_entity_type()
+    pbwg_account = etl_pbwg.get_account()
     updated_pbwg_beneficiaries = etl_pbwg.get_updated_entity_ids(last_success_task_date)
-    Beneficiary.objects.filter(entity_id__in=updated_pbwg_beneficiaries).delete()
+    Beneficiary.objects.filter(account=pbwg_account, entity_id__in=updated_pbwg_beneficiaries).delete()
     Pbwg().run(updated_pbwg_beneficiaries, entity_type_pbwg_code, task_name)
 
     logger.info(
         f"----------------------------- Aggregating PBWG journey for {pbwg_account} per org unit, admission and period(month and year) -----------------------------"
     )
-    MonthlyStatistics.objects.filter(
-        account=pbwg_account,
-        programme_type="PLW",
-        org_unit_id__in=org_units_with_updated_data,
-    ).delete()
-    etl.journey_with_visit_and_steps_per_visit(pbwg_account, "PLW", org_units_with_updated_data)
+    pbwg_org_units = etl_pbwg.get_org_unit_and_period_with_updated_data(last_success_task_date)
+    Aggregator.reset_monthly_statistics(pbwg_account, "PLW", pbwg_org_units)
+    Aggregator.aggregate_monthly_data_by_org_unit(pbwg_account, pbwg_org_units, "PLW")
