+++ plugins/wfp/templates/delete_beneficiaries_analytics.html
@@ -1,6 +1,6 @@
 <html>
 <body>
-<h1>Would you like to delete all beneficiaries?</h1>
+<h1>Would you like to delete all beneficiaries for account {{ account_id }}?</h1>
 
 <form method="post">
     {% csrf_token %}
