+++ plugins/wfp/tests/fixtures/monthly_PBWG_data.json
@@ -1,0 +1,78 @@
+[
+    {
+        "org_unit_id": 9810,
+        "dhis2_id": "OU_DHIS2_ID",
+        "gender": "Male",
+        "month": "8",
+        "year": "2025",
+        "period": "202508",
+        "nutrition_programme": "TSFP",
+        "physiology_status": "Pregnant",
+        "muac_under_11_5": 0,
+        "muac_11_5_12_4": 0,
+        "muac_above_12_5": 0,
+        "muac_under_23": 1,
+        "muac_above_23": 1,
+        "whz_score_2": 0,
+        "whz_score_3": 0,
+        "whz_score_3_2": 0,
+        "oedema_cases": null,
+        "community_health_worker_muac_under_11_5": 0,
+        "community_health_worker_muac_11_5_12_4": 12,
+        "community_health_worker_oedema": 0,
+        "community_health_worker_muac_under_23": 0,
+        "community_health_worker_muac_above_23": 1,
+        "admission_type_new_case": 1,
+        "admission_type_relapse": 0,
+        "admission_type_returned_defaulter": 0,
+        "admission_type_returned_referral": 0,
+        "admission_type_transfer_from_other_tsfp": 0,
+        "admission_type_admission_sc_itp_otp": 8,
+        "admission_type_transfer_sc_itp_otp": 0,
+        "exit_type_transfer_in_from_other_tsfp": 0,
+        "exit_type_cured": 1,
+        "exit_type_death": 0,
+        "exit_type_defaulter": 0,
+        "exit_type_non_respondent": 0,
+        "total_beneficiaries": 1,
+        "number_visits": 3
+    },
+    {
+        "org_unit_id": 9810,
+        "dhis2_id": "OU_DHIS2_ID",
+        "gender": "Male",
+        "month": "8",
+        "year": "2025",
+        "period": "202508",
+        "nutrition_programme": "TSFP",
+        "physiology_status": "Breastfeeding",
+        "muac_under_11_5": null,
+        "muac_11_5_12_4": null,
+        "muac_above_12_5": null,
+        "muac_under_23": 3,
+        "muac_above_23": 2,
+        "whz_score_2": 0,
+        "whz_score_3": 0,
+        "whz_score_3_2": 0,
+        "oedema_cases": null,
+        "community_health_worker_muac_under_11_5": null,
+        "community_health_worker_muac_11_5_12_4": null,
+        "community_health_worker_oedema": null,
+        "community_health_worker_muac_under_23": 3,
+        "community_health_worker_muac_above_23": 2,
+        "admission_type_new_case": 1,
+        "admission_type_relapse": 0,
+        "admission_type_returned_defaulter": 0,
+        "admission_type_returned_referral": 0,
+        "admission_type_transfer_from_other_tsfp": 0,
+        "admission_type_admission_sc_itp_otp": 8,
+        "admission_type_transfer_sc_itp_otp": 0,
+        "exit_type_transfer_in_from_other_tsfp": 0,
+        "exit_type_cured": 1,
+        "exit_type_death": 0,
+        "exit_type_defaulter": 0,
+        "exit_type_non_respondent": 0,
+        "total_beneficiaries": 3,
+        "number_visits": 7
+    }
+]
