--- plugins/wfp/tests/fixtures/monthly_etl_data.json
@@ -1,81 +1,0 @@
-[
-    {
-        "gender": "Male",
-        "month": "8",
-        "year": "2025",
-        "period": "202508",
-        "programme_type": "U5",
-        "nutrition_programme": "TSFP",
-        "admission_type": "new_case",
-        "admission_criteria": "muac",
-        "number_visits": 3,
-        "beneficiary_with_admission_type": 1,
-        "given_sachet_rusf": null,
-        "given_sachet_rutf": null,
-        "given_quantity_csb": null,
-        "given_ration_cbt": null,
-        "muac_under_11_5": 1,
-        "muac_11_5_12_4": 2,
-        "muac_above_12_5": 1,
-        "muac_under_23": null,
-        "muac_above_23": null,
-        "whz_score_2": null,
-        "whz_score_3": null,
-        "whz_score_3_2": null,
-        "oedema": null
-    },
-    {
-        "gender": "Male",
-        "month": "9",
-        "year": "2025",
-        "period": "202509",
-        "programme_type": "U5",
-        "nutrition_programme": "TSFP",
-        "admission_type": "new_case",
-        "admission_criteria": "muac",
-        "number_visits": 1,
-        "beneficiary_with_admission_type": null,
-        "exit_type": "cured",
-        "beneficiary_with_exit_type": 1,
-        "given_sachet_rusf": null,
-        "given_sachet_rutf": null,
-        "given_quantity_csb": null,
-        "given_ration_cbt": null,
-        "muac_under_11_5": null,
-        "muac_11_5_12_4": null,
-        "muac_above_12_5": 1,
-        "muac_under_23": null,
-        "muac_above_23": null,
-        "whz_score_2": null,
-        "whz_score_3": null,
-        "whz_score_3_2": null,
-        "oedema": null
-    },
-    {
-        "gender": "Female",
-        "month": "8",
-        "year": "2025",
-        "period": "202508",
-        "programme_type": "U5",
-        "nutrition_programme": "OTP",
-        "admission_type": "new_case",
-        "admission_criteria": "muac",
-        "number_visits": 3,
-        "beneficiary_with_admission_type": 1,
-        "exit_type": "cured",
-        "beneficiary_with_exit_type": 1,
-        "given_sachet_rusf": null,
-        "given_sachet_rutf": null,
-        "given_quantity_csb": null,
-        "given_ration_cbt": null,
-        "muac_under_11_5": null,
-        "muac_11_5_12_4": 2,
-        "muac_above_12_5": 1,
-        "muac_under_23": null,
-        "muac_above_23": null,
-        "whz_score_2": null,
-        "whz_score_3": null,
-        "whz_score_3_2": null,
-        "oedema": null
-    }
-]
